% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{effects_plot}
\alias{effects_plot}
\title{Plotting}
\usage{
effects_plot(s1v, focal_effects, marker_range = NULL, map = NULL)
}
\arguments{
\item{s1v}{the scan1var object}

\item{focal_effects}{the effects to plot}

\item{marker_range}{the range of markers, defaults to NULL, plotting effects of NULL model}

\item{map}{}
}
\value{
the plot
}
\description{
Plotting
}
