% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1var.R
\name{is_scan1var}
\alias{is_scan1var}
\title{Test whether an R object is a scan1var}
\usage{
is_scan1var(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
TRUE if [x] is a scan1var, FALSE otherwise
}
\description{
Test whether an R object is a scan1var
}
