% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.scan1var}
\alias{plot.scan1var}
\title{Plot scan1var}
\usage{
\method{plot}{scan1var}(x, genetic_map)
}
\arguments{
\item{genetic_map}{the map giving he location of each marker}

\item{s1v}{the scan1var object to be plotted}
}
\value{
the plot
}
\description{
Plot scan1var
}
