% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_allele_effects}
\alias{plot_allele_effects}
\title{Plot allele effects}
\usage{
plot_allele_effects(s1v, marker)
}
\arguments{
\item{s1v}{the scan1var object with the results to be plotted}

\item{marker}{the marker at which the allele effects will be plotted}
}
\value{
the plot
}
\description{
Plot allele effects
}
