% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_allele_effects_over_region}
\alias{plot_allele_effects_over_region}
\title{plot allele effects over a region}
\usage{
plot_allele_effects_over_region(s1v, cross, chr = NULL)
}
\arguments{
\item{s1v}{temp}

\item{cross}{temp}

\item{chr}{temp}
}
\value{
the plot
}
\description{
plot allele effects over a region
}
