% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_allele_effects_scan}
\alias{plot_allele_effects_scan}
\title{plot allele effects over a region}
\usage{
plot_allele_effects_scan(s1v, genetic_map, chr = NULL)
}
\arguments{
\item{s1v}{temp}

\item{genetic_map}{temp}

\item{chr}{temp}
}
\value{
the plot
}
\description{
plot allele effects over a region
}
