% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scan1var}
\alias{plot_scan1var}
\title{Plot scan1var}
\usage{
plot_scan1var(s1v, cross)
}
\arguments{
\item{s1v}{the scan1var object to be plotted}

\item{cross}{the cross object}
}
\value{
the plot
}
\description{
Plot scan1var
}
