% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_DO_ap}
\alias{tiny_DO_ap}
\title{Allele probabilities for the tiny DO cross}
\format{a list of allele probablities as described at
https://kbroman.org/qtl2/assets/vignettes/user_guide.html#calculating_genotype_probabilities
and https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}
\source{
\url{https://kbroman.org/qtl2/pages/sampledata.html}
}
\usage{
tiny_DO_ap
}
\description{
Allele probabilities for the mice from
Gatti et al. (2014) <doi:10.1534/genetics.114.161844>
}
\keyword{datasets}
