% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_DO_cross}
\alias{tiny_DO_cross}
\title{A tiny DO "cross"}
\format{a 'cross' object as described by R package qtl2 and
detailed at https://kbroman.org/qtl2/assets/vignettes/user_guide.html#data_file_format
and https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}
\source{
\url{https://kbroman.org/qtl2/pages/sampledata.html}
}
\usage{
tiny_DO_cross
}
\description{
DO experiment in which white blood cell concentration (WBC)
and neutrophil concentration (NEUT) were measured, as described in
Gatti et al. (2014) <doi:10.1534/genetics.114.161844>
}
\keyword{datasets}
