% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_DO_s1v}
\alias{tiny_DO_s1v}
\title{A tiny scan1var result from a DO experiment}
\format{a data.frame as described...todo}
\source{
\url{https://kbroman.org/qtl2/pages/sampledata.html}
}
\usage{
tiny_DO_s1v
}
\description{
The result of running scan1var on 'tiny_DO_cross' and 'tiny_DO_'ap'
}
\keyword{datasets}
