% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_F2_ap}
\alias{tiny_F2_ap}
\title{Allele probabilities for the tiny F2 cross}
\format{a list of allele probablities as described at
https://kbroman.org/qtl2/assets/vignettes/user_guide.html#calculating_genotype_probabilities
and https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}
\source{
\url{https://kbroman.org/qtl2/pages/sampledata.html}
}
\usage{
tiny_F2_ap
}
\description{
Allele probabilities for the mice from
Grant et al. (2006) <doi:10.1002/hep.21233>
}
\keyword{datasets}
