% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_F2_cross}
\alias{tiny_F2_cross}
\title{A tiny F2 cross}
\format{a 'cross' object as described by R package qtl2 and
detailed at https://kbroman.org/qtl2/assets/vignettes/user_guide.html#data_file_format
and https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}
\source{
\url{https://kbroman.org/qtl2/pages/sampledata.html}
}
\usage{
tiny_F2_cross
}
\description{
F2 intercross studying iron status in mice as described in
Grant et al. (2006) <doi:10.1002/hep.21233>
}
\keyword{datasets}
