% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny_F2_s1v}
\alias{tiny_F2_s1v}
\title{A tiny scan1var result from an F2 cross}
\format{a data.frame as described...todo}
\source{
\url{https://kbroman.org/qtl2/pages/sampledata.html}
}
\usage{
tiny_F2_s1v
}
\description{
The result of running scan1var on 'tiny_F2_cross' and 'tiny_F2_'ap'
}
\keyword{datasets}
