% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan1var.R
\name{scan1var}
\alias{scan1var}
\title{Mean-Variance genome scan with a single-QTL model}
\usage{
scan1var(pheno_name, mean_covar_names = NULL, var_covar_names = NULL,
  alleleprobs, non_genetic_data, model = c("normal", "binary"),
  weights = NULL, num_cores = 1, ...)
}
\arguments{
\item{alleleprobs}{Genotype probabilities as calculated by
[qtl::calc_genoprob()].}

\item{model}{Indicates whether to use a normal model (least
squares) or binary model (logistic regression) for the phenotype.
If `model="binary"`, the phenotypes must have values in \eqn{[0, 1]}.}

\item{weights}{An optional numeric vector of positive weights for the
individuals. As with the other inputs, it must have `names`
for individual identifiers.}

\item{num_cores}{Number of CPU cores to use, for parallel calculations.
(If `0`, use [parallel::detectCores() - 1].)}

\item{...}{additional optional arguments}

\item{pheno}{A data.frame of phenotypes, individuals x phenotypes.}

\item{mean_addcovar}{An optional data.frame of additive covariates.}

\item{var_addcovar}{An optional data.frame of additive covariates.}
}
\value{
results of the scan
}
\description{
Mean-Variance genome scan with a single-QTL model
}
